+++ src/egon/data/datasets/heat_demand_timeseries/HTS.py
@@ -9,6 +9,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 import pandas.io.sql as sqlio
+from pathlib import Path
 import psycopg2
 import xarray as xr
 
@@ -998,14 +999,14 @@ def profile_generator(aggregation_level):
 
     heat_profile_dist : pandas.DataFrame
         if aggreation_level = 'district'
-            heat profiles for every distric heating id
+            normalized heat profiles for every distric heating id
         else
             0
     heat_profile_idp : pandas.DataFrame
         if aggreation_level = 'district'
-            heat profiles for every mv grid bus_id
+            normalized heat profiles for every mv grid bus_id
         else
-            heat profiles for every zensus_poppulation_id
+            normalized heat profiles for every zensus_poppulation_id
     """
 
     annual_demand, idp_df, selected_profiles = profile_selector()
@@ -1076,6 +1077,7 @@ def profile_generator(aggregation_level):
                 right_on=mv_grid_ind.index,
                 how="inner",
             )
+
             heat_profile_idp.sort_values("bus_id", inplace=True)
             heat_profile_idp.set_index("bus_id", inplace=True)
             heat_profile_idp.drop("key_0", axis=1, inplace=True)
@@ -1159,20 +1161,7 @@ def residential_demand_scale(aggregation_level):
     h = h_value()
     h = h.reset_index(drop=True)
 
-    district_heating = psycop_df_AF(
-        "demand.egon_map_zensus_district_heating_areas"
-    )
-
-    district_heating = district_heating.pivot_table(
-        values="area_id", index="zensus_population_id", columns="scenario"
-    )
-
     mv_grid = psycop_df_AF("boundaries.egon_map_zensus_grid_districts")
-
-    mv_grid_ind = mv_grid.loc[
-        mv_grid.index.difference(district_heating.index), :
-    ]
-    mv_grid_ind = mv_grid_ind.reset_index()
 
     scenarios = ["eGon2035", "eGon100RE"]
 
@@ -1181,6 +1170,25 @@ def residential_demand_scale(aggregation_level):
     residential_zensus_profile = pd.DataFrame()
 
     for scenario in scenarios:
+
+        district_heating = db.select_dataframe(
+            f"""
+            SELECT * FROM 
+            demand.egon_map_zensus_district_heating_areas
+            WHERE scenario = '{scenario}'
+            """
+        )
+
+        district_heating = district_heating.pivot_table(
+            values="area_id", index="zensus_population_id", columns="scenario"
+        )
+
+        mv_grid_ind = mv_grid[
+            ~mv_grid.zensus_population_id.isin(district_heating.index)
+        ]
+
+        mv_grid_ind = mv_grid_ind.reset_index()
+
         if aggregation_level == "district":
 
             scenario_ids = district_heating[scenario]
@@ -1264,6 +1272,7 @@ def residential_demand_scale(aggregation_level):
                 heat_demand_profile_dist
             )
 
+            # Individual supplied heat demand time series
             district_grid = pd.merge(
                 mv_grid_ind[["bus_id", "zensus_population_id"]],
                 annual_demand[["zensus_population_id", "Station", scenario]],
@@ -1721,6 +1730,57 @@ def CTS_demand_scale(aggregation_level):
     return CTS_district, CTS_grid, CTS_zensus
 
 
+def store_national_profiles(
+    residential_demand_grid,
+    CTS_demand_grid,
+    residential_demand_dist,
+    CTS_demand_dist,
+):
+
+    folder = Path(".") / "input-pypsa-eur-sec"
+    # Create the folder, if it does not exists already
+    if not os.path.exists(folder):
+        os.mkdir(folder)
+
+    for scenario in CTS_demand_grid.scenario.unique():
+        national_demand = pd.DataFrame(
+            columns=["residential rural", "services rural", "urban central"],
+            index=pd.date_range(
+                datetime(2011, 1, 1, 0), periods=8760, freq="H"
+            ),
+        )
+
+        national_demand["residential rural"] = (
+            residential_demand_grid[
+                residential_demand_grid.scenario == scenario
+            ]
+            .drop("scenario", axis="columns")
+            .sum()
+            .values
+        )
+        national_demand["services rural"] = (
+            CTS_demand_grid[CTS_demand_grid.scenario == scenario]
+            .sum(numeric_only=True)
+            .values
+        )
+        national_demand["urban central"] = (
+            residential_demand_dist[
+                residential_demand_dist.scenario == scenario
+            ]
+            .drop("scenario", axis="columns")
+            .sum()
+            .values
+            + CTS_demand_dist[CTS_demand_dist.scenario == scenario]
+            .drop("scenario", axis="columns")
+            .sum()
+            .values
+        )
+
+        national_demand.to_csv(
+            folder / f"heat_demand_timeseries_DE_{scenario}.csv"
+        )
+
+
 def demand_profile_generator(aggregation_level="district"):
     """
 
@@ -1745,9 +1805,68 @@ def demand_profile_generator(aggregation_level="district"):
         residential_demand_zensus,
     ) = residential_demand_scale(aggregation_level)
 
+    # Compare with target value
+    target = db.select_dataframe(
+        """
+        SELECT scenario, SUM(demand) as demand
+        FROM demand.egon_peta_heat
+        WHERE sector = 'residential'
+        GROUP BY (scenario)
+        """,
+        index_col="scenario",
+    )
+
+    check_residential = (
+        (
+            residential_demand_dist.groupby("scenario").sum().sum(axis=1)
+            + residential_demand_grid.groupby("scenario").sum().sum(axis=1)
+        )
+        - target.demand
+    ) / target.demand
+
+    assert (
+        check_residential.abs().max() < 0.01
+    ), f"""Unexpected deviation between target value and distributed
+        residential heat demand: {check_residential}
+        """
+
     CTS_demand_dist, CTS_demand_grid, CTS_demand_zensus = CTS_demand_scale(
         aggregation_level
     )
+
+    # Compare with target value
+    target_cts = db.select_dataframe(
+        """
+        SELECT scenario, SUM(demand) as demand
+        FROM demand.egon_peta_heat
+        WHERE sector = 'service'
+        GROUP BY (scenario)
+        """,
+        index_col="scenario",
+    )
+
+    check_cts = (
+        (
+            CTS_demand_dist.groupby("scenario").sum().sum(axis=1)
+            + CTS_demand_grid.groupby("scenario").sum().sum(axis=1)
+        )
+        - target_cts.demand
+    ) / target_cts.demand
+
+    assert (
+        check_cts.abs().max() < 0.01
+    ), f"""Unexpected deviation between target value and distributed
+        service heat demand: {check_residential}
+        """
+
+    # store demand timeseries for pypsa-eur-sec on national level
+    store_national_profiles(
+        residential_demand_grid,
+        CTS_demand_grid,
+        residential_demand_dist,
+        CTS_demand_dist,
+    )
+
 
     if aggregation_level == "district":
         total_demands_dist = pd.concat(
@@ -1792,6 +1911,7 @@ def demand_profile_generator(aggregation_level="district"):
 
         final_heat_profiles_grid = pd.DataFrame()
         for scenario in scenarios:
+
             scenario_demand = (
                 total_demands_grid[total_demands_grid.scenario == scenario]
                 .drop("scenario", axis=1)
@@ -1862,7 +1982,7 @@ class HeatTimeSeries(Dataset):
     def __init__(self, dependencies):
         super().__init__(
             name="HeatTimeSeries",
-            version="0.0.5",
+            version="0.0.7",
             dependencies=dependencies,
             tasks=(demand_profile_generator),
         )