+++ src/egon/data/datasets/pypsaeursec/__init__.py
@@ -16,9 +16,9 @@ import yaml
 
 from egon.data import __path__, db
 from egon.data.datasets import Dataset
+from egon.data.datasets.scenario_parameters import get_sector_parameters
 import egon.data.config
 import egon.data.subprocess as subproc
-from egon.data.datasets.scenario_parameters import get_sector_parameters
 
 
 def run_pypsa_eur_sec():
@@ -28,6 +28,7 @@ def run_pypsa_eur_sec():
     filepath.mkdir(parents=True, exist_ok=True)
 
     pypsa_eur_repos = filepath / "pypsa-eur"
+    pypsa_eur_repos_data = pypsa_eur_repos / "data"
     technology_data_repos = filepath / "technology-data"
     pypsa_eur_sec_repos = filepath / "pypsa-eur-sec"
     pypsa_eur_sec_repos_data = pypsa_eur_sec_repos / "data"
@@ -38,16 +39,16 @@ def run_pypsa_eur_sec():
                 "git",
                 "clone",
                 "--branch",
-                "master",
+                "v0.4.0",
                 "https://github.com/PyPSA/pypsa-eur.git",
                 pypsa_eur_repos,
             ]
         )
 
-        subproc.run(
-            ["git", "checkout", "4e44822514755cdd0289687556547100fba6218b"],
-            cwd=pypsa_eur_repos,
-        )
+        # subproc.run(
+        #     ["git", "checkout", "4e44822514755cdd0289687556547100fba6218b"],
+        #     cwd=pypsa_eur_repos,
+        # )
 
         file_to_copy = os.path.join(
             __path__[0], "datasets", "pypsaeursec", "pypsaeur", "Snakefile"
@@ -68,13 +69,22 @@ def run_pypsa_eur_sec():
                 env, outfile, default_flow_style=False, allow_unicode=True
             )
 
+        datafile = "pypsa-eur-data-bundle.tar.xz"
+        datapath = pypsa_eur_repos / datafile
+        if not datapath.exists():
+            urlretrieve(
+                f"https://zenodo.org/record/3517935/files/{datafile}", datapath
+            )
+            tar = tarfile.open(datapath)
+            tar.extractall(pypsa_eur_repos_data)
+
     if not technology_data_repos.exists():
         subproc.run(
             [
                 "git",
                 "clone",
                 "--branch",
-                "v0.2.0",
+                "v0.3.0",
                 "https://github.com/PyPSA/technology-data.git",
                 technology_data_repos,
             ]
@@ -90,10 +100,12 @@ def run_pypsa_eur_sec():
             ]
         )
 
-    datafile = "pypsa-eur-sec-data-bundle-210418.tar.gz"
+    datafile = "pypsa-eur-sec-data-bundle.tar.gz"
     datapath = pypsa_eur_sec_repos_data / datafile
     if not datapath.exists():
-        urlretrieve(f"https://nworbmot.org/{datafile}", datapath)
+        urlretrieve(
+            f"https://zenodo.org/record/5824485/files/{datafile}", datapath
+        )
         tar = tarfile.open(datapath)
         tar.extractall(pypsa_eur_sec_repos_data)
 
@@ -117,7 +129,7 @@ def run_pypsa_eur_sec():
     )
 
 
-def neighbor_reduction():
+def read_network():
 
     # Set execute_pypsa_eur_sec to False until optional task is implemented
     execute_pypsa_eur_sec = False
@@ -164,7 +176,12 @@ def neighbor_reduction():
             / "elec_s_37_lv2.0__Co2L0-3H-T-H-B-I-dist1_2050.nc"
         )
 
-    network = pypsa.Network(str(target_file))
+    return pypsa.Network(str(target_file))
+
+
+def neighbor_reduction():
+
+    network = read_network()
 
     wanted_countries = [
         "DE",
@@ -387,8 +404,11 @@ def neighbor_reduction():
     neighbor_gens = network.generators[
         network.generators.bus.isin(neighbors.index)
     ]
-    neighbor_gens_t = network.generators_t["p_max_pu"][neighbor_gens[neighbor_gens.index.isin(network.generators_t["p_max_pu"].columns)].index]
-
+    neighbor_gens_t = network.generators_t["p_max_pu"][
+        neighbor_gens[
+            neighbor_gens.index.isin(network.generators_t["p_max_pu"].columns)
+        ].index
+    ]
 
     neighbor_gens.reset_index(inplace=True)
     neighbor_gens.bus = (
@@ -461,12 +481,33 @@ def neighbor_reduction():
     # Connect to local database
     engine = db.engine()
 
-    # db.execute_sql("DELETE FROM grid.egon_etrago_bus "
-    #               "WHERE scn_name = 'eGon100RE' "
-    #               "AND country <> 'DE'")
+    db.execute_sql(
+        "DELETE FROM grid.egon_etrago_bus "
+        "WHERE scn_name = 'eGon100RE' "
+        "AND country <> 'DE'"
+    )
 
     neighbors["scn_name"] = "eGon100RE"
     neighbors.index = neighbors["new_index"]
+
+    # Correct geometry for non AC buses
+    carriers = set(neighbors.carrier.to_list())
+    carriers.remove("AC")
+    non_AC_neighbors = pd.DataFrame()
+    for c in carriers:
+        c_neighbors = neighbors[neighbors.carrier == c].set_index(
+            "location", drop=False
+        )
+        for i in ["x", "y"]:
+            c_neighbors = c_neighbors.drop(i, axis=1)
+        coordinates = neighbors[neighbors.carrier == "AC"][
+            ["location", "x", "y"]
+        ].set_index("location")
+        c_neighbors = pd.concat([coordinates, c_neighbors], axis=1).set_index(
+            "new_index", drop=False
+        )
+        non_AC_neighbors = non_AC_neighbors.append(c_neighbors)
+    neighbors = neighbors[neighbors.carrier == "AC"].append(non_AC_neighbors)
 
     for i in ["new_index", "control", "generator", "location", "sub_network"]:
         neighbors = neighbors.drop(i, axis=1)
@@ -817,7 +858,7 @@ class PypsaEurSec(Dataset):
     def __init__(self, dependencies):
         super().__init__(
             name="PypsaEurSec",
-            version="0.0.4",
+            version="0.0.6",
             dependencies=dependencies,
             tasks=tasks,
         )